Add tests for LocalTime.
Move tests for local-time to new file.
Add tests for errors in setters.
